Remove setting selection on every onSlateChange
